Update user objects in `server.js`
The commit introduces modifications to the `server.js` file. The changes include updating the user objects by adding a new property `role` with corresponding values for both `admin` and `client` users.

In addition, the commit adds a check to verify if the request is authenticated. If not, it sends a response with a status code of 401 and a message of 'Unauthorized'.